Update: Added default server and clean up

+++ src/utils.rs
@@ -1,5 +1,6 @@
 use anyhow::Context;
 use anyhow::{anyhow, Result};
+use async_recursion::async_recursion;
 use env_logger::fmt::Color;
 use log::Level;
 use std::collections::HashMap;
@@ -8,21 +9,30 @@ use std::fs::remove_file;
 use std::io::ErrorKind;
 use std::io::Write;
 use std::os::unix::fs::symlink;
+use std::path::Path;
 use std::process::Command;
+use tokio::fs::{create_dir_all, File};
+use tokio::io::AsyncReadExt;
 use walkdir::WalkDir;
 
 pub const AVAILABLE: &str = "sites-available";
 pub const ENABLED: &str = "sites-enabled";
 
+#[derive(Debug)]
 pub struct FileData {
     pub file_name: String,
     pub is_symlink: bool,
 }
 
-pub fn walk_folder(folder: &str) -> Result<HashMap<String, FileData>> {
+pub async fn walk_folder(folder: &str) -> Result<HashMap<String, FileData>> {
     let mut files: HashMap<String, FileData> = HashMap::new();
-    let folder_path = format!("/etc/nginx/{}", folder);
-    let iter = WalkDir::new(&folder_path).max_depth(1);
+    let dir = Path::new("/etc/nginx").join(folder);
+
+    if !dir.exists() {
+        create_dir_all(&dir).await?;
+    }
+
+    let iter = WalkDir::new(&dir).max_depth(1);
     for entry in iter {
         let entry = entry?;
         let file_name: String = entry.file_name().to_string_lossy().into();
@@ -45,18 +55,34 @@ pub fn walk_folder(folder: &str) -> Result<HashMap<String, FileData>> {
     Ok(files)
 }
 
-pub fn sym_link(file: String) -> Result<()> {
-    let available_path = format!("/etc/nginx/{}/{}", AVAILABLE, file);
-    let enabled_path = format!("/etc/nginx/{}/{}", ENABLED, file);
+#[async_recursion]
+pub async fn sym_link(file: String) -> Result<()> {
+    let available_dir = Path::new("/etc/nginx").join(AVAILABLE);
+    let enabled_dir = Path::new("/etc/nginx").join(ENABLED);
 
-    let symlink_res = symlink(available_path, enabled_path);
+    if !enabled_dir.exists() {
+        create_dir_all(&enabled_dir).await?;
+    }
+
+    if !available_dir.exists() {
+        create_dir_all(&available_dir).await?;
+    }
+
+    let enabled_path = enabled_dir.join(&file);
+    let available_path = available_dir.join(&file);
+
+    if !available_path.exists() {
+        log::error!("Failed to symlink, file not found");
+
+        return Ok(());
+    }
+
+    let symlink_res = symlink(&available_path, &enabled_path);
 
     if let Err(err) = symlink_res {
         if err.kind() == ErrorKind::AlreadyExists {
-            let enabled_path = format!("/etc/nginx/{}/{}", ENABLED, file);
-
             remove_file(enabled_path)?;
-            return sym_link(file);
+            return sym_link(file).await;
         }
 
         log::error!("Failed to symlink");
@@ -66,9 +92,14 @@ pub fn sym_link(file: String) -> Result<()> {
     Ok(())
 }
 
-pub fn rm_symlink(file: String) -> Result<()> {
-    let enabled_path = format!("/etc/nginx/{}/{}", ENABLED, file);
-    remove_file(enabled_path)?;
+pub async fn rm_symlink(file_name: String) -> Result<()> {
+    let enabled_dir = Path::new("/etc/nginx").join(ENABLED);
+    let file_path = enabled_dir.join(file_name);
+
+    if file_path.exists() {
+        remove_file(file_path)?;
+    }
+
     Ok(())
 }
 
@@ -107,7 +138,14 @@ pub fn reload_nginx() -> Result<()> {
 }
 
 pub fn edit_nginx_site(file_name: String) -> Result<()> {
-    let file_path = format!("/etc/nginx/{}/{}", AVAILABLE, file_name);
+    let available_dir = Path::new("/etc/nginx").join(AVAILABLE);
+    let file_path = available_dir.join(file_name);
+
+    if !file_path.exists() {
+        log::info!("File not found.");
+
+        return Ok(());
+    }
 
     let output = Command::new("vi")
         .arg(file_path)
@@ -122,6 +160,27 @@ pub fn edit_nginx_site(file_name: String) -> Result<()> {
     }
 
     log::info!("Edit done.");
+    Ok(())
+}
+
+pub async fn view_nginx_site(file_name: String) -> Result<()> {
+    let available_dir = Path::new("/etc/nginx").join(AVAILABLE);
+    let file_path = available_dir.join(&file_name);
+
+    if !file_path.exists() {
+        log::info!("File not found.");
+
+        return Ok(());
+    }
+
+    let mut file = File::open(file_path).await?;
+    let mut buffer = String::new();
+
+    file.read_to_string(&mut buffer).await?;
+
+    println!("---------------------- Start of {file_name} ----------------------");
+    println!("{buffer}");
+    println!("---------------------- End of {file_name} ----------------------");
     Ok(())
 }
